Add a base DeleteView class
